Shader reformatting

+++ Terrain Visualizer/include/camera.h
@@ -100,7 +100,7 @@ struct camera {
 	
 	inline void move(float3 v, float2 d) {
 		rot += d;
-		const float k = -cos(rot.y);
+		const float k = -cosf(rot.y);
 		const float u = (k < 0) ? 1.0f : -1.0f;
 		const float3 forward = norm(float3{k * sinf(rot.x), sinf(rot.y), k * cosf(rot.x)});
 		const float3 right = norm(float3{-forward.z, 0.0f, forward.x}) * u;
